orders_service/requirements.txt
Flask
PyJWT==2.8.0